+++ src/main/java/net/minestom/server/instance/light/LightCompute.java
@@ -71,8 +71,8 @@ public final class LightCompute {
                 // Section
                 final int newIndex = xO | (zO << 4) | (yO << 8);
                 if (getLight(lightArray, newIndex) + 2 <= lightLevel) {
-                    final Block currentBlock = Objects.requireNonNullElse(Block.fromStateId((short)blockPalette.get(x, y, z)), Block.AIR);
-                    final Block propagatedBlock = Objects.requireNonNullElse(Block.fromStateId((short)blockPalette.get(xO, yO, zO)), Block.AIR);
+                    final Block currentBlock = Objects.requireNonNullElse(Block.fromStateId(blockPalette.get(x, y, z)), Block.AIR);
+                    final Block propagatedBlock = Objects.requireNonNullElse(Block.fromStateId(blockPalette.get(xO, yO, zO)), Block.AIR);
 
                     boolean airAir = currentBlock.isAir() && propagatedBlock.isAir();
                     if (!airAir && currentBlock.registry().collisionShape().isOccluded(propagatedBlock.registry().collisionShape(), face)) continue;
